Merge: Merging with master and refactoring
